refactor: StagingArea 및 TargetArea 관련 interface 정리 및 분리
