--- src/types/today/TargetControlSectionProps.ts
@@ -1,7 +1,0 @@
-export interface TargetControlSectionProps {
-	onClickPrevDate: () => void;
-	onClickNextDate: () => void;
-	onClickTodayDate: () => void;
-	onClickDatePicker: (target: Date) => void;
-	targetDate: string;
-}
